fix(singleproduct): edit files single product
